+++ src/main/java/VR/entities/EntitySuper.java
@@ -3,8 +3,8 @@ package VR.entities;
 
 
 public class EntitySuper {
-    protected int x;
-    protected int y;
+    protected double x;
+    protected double y;
     
     
     public EntitySuper() {
@@ -12,26 +12,26 @@ public class EntitySuper {
         this.y = 0;
     }
     
-    public EntitySuper(int x, int y) {
+    public EntitySuper(double x, double y) {
         this.x = x;
         this.y = y;
     }
     
-    public void setXY(int x, int y) {
+    public void setXY(double x, double y) {
         this.x = x;
         this.y = y;
     }
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
     
-    public int getX() {
+    public double getX() {
         return x;
     }
-    public int getY() {
+    public double getY() {
         return y;
     }
